refactor(core): split facade by responsibility
