+++ cli/internal/core/core.go
@@ -1,32 +1,21 @@
 package core
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"io"
-	"maps"
-	"strings"
 	"time"
 
 	"github.com/charmbracelet/log"
 	"github.com/nats-io/nats.go"
 	"github.com/nats-io/nats.go/jetstream"
 	"golang.org/x/sync/errgroup"
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/timestamppb"
-	"hmans.de/chatto/internal/assets"
+
 	"hmans.de/chatto/internal/config"
 	"hmans.de/chatto/internal/core/linkpreview"
-	"hmans.de/chatto/internal/dekstore"
 	"hmans.de/chatto/internal/evtstream"
-	"hmans.de/chatto/internal/kms"
 	corev1 "hmans.de/chatto/internal/pb/chatto/core/v1"
-	"hmans.de/chatto/pkg/events"
 )
-
-const projectionSnapshotObjectStoreName = "PROJECTION_SNAPSHOTS"
 
 // ============================================================================
 // ChattoCore
@@ -290,30 +279,8 @@ func (c *ChattoCore) EventStreamForDebug(_ context.Context) (jetstream.Stream, e
 	return c.storage.serverEvtStream, nil
 }
 
-// assetURL prepends AssetBaseURL to an asset path.
-// When AssetBaseURL is empty, returns the path unchanged.
-func (c *ChattoCore) assetURL(path string) string {
-	if c.AssetBaseURL == "" {
-		return path
-	}
-	return c.AssetBaseURL + path
-}
-
-// encryptionManager handles message body encryption/decryption.
-type encryptionManager struct {
-	keyWrapper  kms.KeyWrapper
-	legacyKeys  kms.LegacyKeyProvider
-	callKeys    kms.CallKeyStore
-	contentKeys *dekstore.Store
-}
-
 func (c *ChattoCore) ServerStore() jetstream.ObjectStore {
 	return c.storage.serverAssets
-}
-
-// KeyWrapper returns the key-only KMS boundary used by encryption operations.
-func (c *ChattoCore) KeyWrapper() kms.KeyWrapper {
-	return c.encryption.keyWrapper
 }
 
 // ConfigModel returns the runtime configuration model.
@@ -325,686 +292,6 @@ func (c *ChattoCore) ConfigModel() *ConfigModel {
 // This implements the server < space < room specificity model.
 func (c *ChattoCore) PermResolver() *PermissionResolver {
 	return c.permissionResolver
-}
-
-// DeleteUserEncryptionKey permanently deletes a user's encryption key (crypto-shredding).
-// All messages encrypted with this key become permanently unreadable.
-// This is used for GDPR-compliant user deletion.
-func (c *ChattoCore) DeleteUserEncryptionKey(ctx context.Context, userID string) error {
-	return c.DeleteUserEncryptionKeyAs(ctx, userID, userID)
-}
-
-func (c *ChattoCore) deleteEncryptionKeyOnly(ctx context.Context, keyRef string) error {
-	if c.encryption.keyWrapper == nil {
-		return nil
-	}
-	return c.encryption.keyWrapper.ShredKey(ctx, keyRef)
-}
-
-func (c *ChattoCore) DeleteUserEncryptionKeyAs(ctx context.Context, actorID, userID string) error {
-	if c.encryption.keyWrapper == nil {
-		return nil // Encryption not configured
-	}
-
-	if err := c.userModel.waitForContentKeysCurrent(ctx, userID); err != nil {
-		return err
-	}
-
-	contentKeyRefs, wrappingKeyRefs, err := c.userModel.keyRefsForShredding(userID)
-	if err != nil {
-		return err
-	}
-	keyRefs := make(map[string]struct{})
-	keyRefs[kms.LegacyUserKeyRef(userID)] = struct{}{}
-	for _, keyRef := range wrappingKeyRefs {
-		if keyRef != "" {
-			keyRefs[keyRef] = struct{}{}
-		}
-	}
-	for _, contentKeyRef := range contentKeyRefs {
-		if c.encryption.contentKeys == nil {
-			return fmt.Errorf("content key store is not configured")
-		}
-		stored, err := c.encryption.contentKeys.Get(ctx, contentKeyRef)
-		if err != nil {
-			return fmt.Errorf("failed to load DEK %s before shredding: %w", contentKeyRef, err)
-		}
-		if wrappingKeyRef := stored.GetWrappingKeyRef(); wrappingKeyRef != "" {
-			keyRefs[wrappingKeyRef] = struct{}{}
-		}
-	}
-
-	shredded := false
-	for _, contentKeyRef := range contentKeyRefs {
-		if err := c.encryption.contentKeys.Shred(ctx, contentKeyRef); err != nil {
-			return err
-		}
-		shredded = true
-	}
-
-	for keyRef := range keyRefs {
-		exists, err := c.encryption.keyWrapper.KeyExists(ctx, keyRef)
-		if err != nil {
-			return err
-		}
-		if !exists {
-			continue
-		}
-		if err := c.encryption.keyWrapper.ShredKey(ctx, keyRef); err != nil {
-			return err
-		}
-		shredded = true
-	}
-	if !shredded {
-		return nil
-	}
-	forgetDEKRequestCacheUser(ctx, userID)
-
-	event := newEvent(actorID, &corev1.Event{
-		Event: &corev1.Event_UserKeyShredded{
-			UserKeyShredded: &corev1.UserKeyShreddedEvent{UserId: userID},
-		},
-	})
-	seq, err := c.appendUserEvent(ctx, userID, event, "", nil)
-	if err != nil {
-		return fmt.Errorf("failed to record user key shred event: %w", err)
-	}
-	subject := evtstream.UserAggregate(userID).SubjectFor(event)
-	return c.roomModel.waitForTimelineAndThreads(ctx, events.SubjectPosition(subject, seq))
-}
-
-// AssetsConfig returns the assets configuration as an assets.Config.
-func (c *ChattoCore) AssetsConfig() assets.Config {
-	maxUploadSize := int64(c.config.Assets.MaxUploadSize)
-	if maxUploadSize == 0 {
-		maxUploadSize = assets.DefaultMaxUploadSize
-	}
-	return assets.Config{
-		MaxUploadSize: maxUploadSize,
-	}
-}
-
-// ShouldUseS3 returns true if new uploads should be stored in S3.
-func (c *ChattoCore) ShouldUseS3() bool {
-	return c.config.Assets.StorageBackend == config.StorageBackendS3 && c.s3Client != nil
-}
-
-// GetLinkPreview fetches link preview metadata for a URL.
-// Results are cached server-side. Returns nil if the URL cannot be previewed.
-func (c *ChattoCore) GetLinkPreview(ctx context.Context, url string) (*corev1.LinkPreview, error) {
-	// Check cache first
-	cached, err := c.linkPreviewCache.Get(ctx, url)
-	if errors.Is(err, linkpreview.ErrCachedFailure) {
-		// Negative cache hit - URL previously failed, don't re-fetch
-		return nil, nil
-	}
-	if err != nil {
-		c.logger.Warn("Failed to get cached link preview", "url", url, "error", err)
-		// Continue to fetch - don't fail on cache errors
-	}
-	if cached != nil {
-		if err := c.markCachedLegacyLinkPreviewPublic(ctx, cached); err != nil {
-			c.logger.Warn("Failed to preserve cached legacy link-preview image", "asset_id", cached.GetImageAssetId(), "error", err)
-		}
-		return cached, nil
-	}
-
-	// Fetch the preview
-	result, err := c.linkPreviewFetcher.Fetch(ctx, url)
-	if err != nil {
-		// Cache the failure to avoid repeated fetches
-		_ = c.linkPreviewCache.SetFailure(ctx, url, err.Error())
-		if errors.Is(err, linkpreview.ErrUnavailable) {
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	preview := result.ToProto(url)
-	if err := validateLinkPreview(preview); err != nil {
-		_ = c.linkPreviewCache.SetFailure(ctx, url, err.Error())
-		c.logger.Warn("Discarding invalid fetched link preview", "url", url, "error", err)
-		return nil, nil
-	}
-
-	// Cache the result
-	if err := c.linkPreviewCache.Set(ctx, url, preview); err != nil {
-		c.logger.Warn("Failed to cache link preview", "url", url, "error", err)
-	}
-
-	return preview, nil
-}
-
-// markCachedLegacyLinkPreviewPublic preserves a pre-namespace link-preview
-// image that is still referenced by the server's runtime cache but has not yet
-// appeared in durable message history. The cached server-issued AssetRecord
-// must bind one exact canonical flat NATS key; private declarations and metadata
-// always win. Only object metadata is updated—the object body is never opened.
-func (c *ChattoCore) markCachedLegacyLinkPreviewPublic(ctx context.Context, preview *corev1.LinkPreview) error {
-	if preview == nil || c.assetModel == nil || c.assetModel.assets.Projection() == nil {
-		return nil
-	}
-	asset := preview.GetImageAsset()
-	assetID := preview.GetImageAssetId()
-	if asset == nil || assetID == "" || asset.GetId() != assetID {
-		return nil
-	}
-	natsAsset := asset.GetNats()
-	if natsAsset == nil || natsAsset.GetKey() != assetID {
-		return nil
-	}
-	logicalID, namespaced, ok := serverAssetRequestKey(natsAsset.GetKey())
-	if !ok || namespaced || logicalID != assetID {
-		return nil
-	}
-	assetState := c.assetModel.AssetState(assetID)
-	if assetState.Creation != nil || assetState.Deleted {
-		return nil
-	}
-
-	info, err := c.storage.serverAssets.GetInfo(ctx, assetID)
-	if err != nil {
-		if errors.Is(err, jetstream.ErrObjectNotFound) {
-			return nil
-		}
-		return fmt.Errorf("inspect legacy preview object metadata: %w", err)
-	}
-	if info.Headers.Get("Room-Id") != "" || info.Headers.Get("Upload-Id") != "" ||
-		info.Headers.Get(ServerAssetVisibilityHeader) == ServerAssetVisibilityPublic {
-		return nil
-	}
-
-	headers := maps.Clone(info.Headers)
-	if headers == nil {
-		headers = make(nats.Header)
-	}
-	headers.Set(ServerAssetVisibilityHeader, ServerAssetVisibilityPublic)
-	headers.Set(ServerAssetVisibilityNUIDHeader, info.NUID)
-	headers.Set(ServerAssetVisibilityDigestHeader, info.Digest)
-	if err := c.storage.serverAssets.UpdateMeta(ctx, assetID, jetstream.ObjectMeta{
-		Name:        assetID,
-		Description: info.Description,
-		Headers:     headers,
-		Metadata:    maps.Clone(info.Metadata),
-	}); err != nil {
-		return fmt.Errorf("mark legacy preview object public: %w", err)
-	}
-	updated, err := c.storage.serverAssets.GetInfo(ctx, assetID)
-	if err != nil {
-		return fmt.Errorf("verify legacy preview object metadata: %w", err)
-	}
-	if updated.Headers.Get("Room-Id") != "" || updated.Headers.Get("Upload-Id") != "" ||
-		!serverAssetVisibilityMarkerMatches(updated) {
-		return fmt.Errorf("legacy preview object generation changed during metadata update")
-	}
-	return nil
-}
-
-// HydrateLinkPreviewImageAsset ensures a posted LinkPreview carries the
-// server-issued AssetRecord for its preview image. Clients only send
-// image_asset_id for compatibility; the backend rehydrates the storage pointer
-// from the server-side preview cache or by probing known backends.
-func (c *ChattoCore) HydrateLinkPreviewImageAsset(ctx context.Context, preview *corev1.LinkPreview) error {
-	if preview == nil {
-		return nil
-	}
-
-	imageAsset := preview.GetImageAsset()
-	imageAssetID := preview.GetImageAssetId()
-	if imageAsset != nil {
-		if imageAsset.GetId() == "" {
-			return fmt.Errorf("link preview image asset record is missing id")
-		}
-		if imageAssetID != "" && imageAssetID != imageAsset.GetId() {
-			return fmt.Errorf("link preview image asset id mismatch")
-		}
-		if imageAssetID == "" {
-			id := imageAsset.GetId()
-			preview.ImageAssetId = &id
-		}
-		return nil
-	}
-	if imageAssetID == "" {
-		return nil
-	}
-
-	if cached, err := c.linkPreviewCache.Get(ctx, preview.GetUrl()); err == nil && cached != nil {
-		if cachedAsset := cached.GetImageAsset(); cachedAsset != nil && cachedAsset.GetId() == imageAssetID {
-			preview.ImageAsset = proto.Clone(cachedAsset).(*corev1.AssetRecord)
-			return nil
-		}
-	} else if err != nil && !errors.Is(err, linkpreview.ErrCachedFailure) {
-		c.logger.Debug("Failed to hydrate link preview image asset from cache", "url", preview.GetUrl(), "error", err)
-	}
-
-	asset, err := c.ServerAssetRecordFromAnyBackend(ctx, imageAssetID, "link-preview.webp")
-	if err != nil {
-		return fmt.Errorf("hydrate link preview image asset: %w", err)
-	}
-	preview.ImageAsset = asset
-	return nil
-}
-
-// S3Client returns the S3 client, or nil if S3 is not configured.
-func (c *ChattoCore) S3Client() *S3Client {
-	return c.s3Client
-}
-
-func (c *ChattoCore) storeLinkPreviewImage(ctx context.Context, assetID string, data []byte, contentType string) (*corev1.AssetRecord, error) {
-	asset := &corev1.AssetRecord{
-		Id:          assetID,
-		Filename:    "link-preview.webp",
-		ContentType: contentType,
-		Size:        int64(len(data)),
-	}
-	if c.ShouldUseS3() {
-		s3Key := S3KeyServerAsset(assetID)
-		if _, err := c.s3Client.PutObjectFromBytes(ctx, s3Key, data, contentType); err != nil {
-			return nil, fmt.Errorf("upload link preview image to S3: %w", err)
-		}
-		asset.Storage = &corev1.AssetRecord_S3{S3: &corev1.S3Asset{
-			Key:    assetID,
-			Bucket: proto.String(c.s3Client.Bucket()),
-		}}
-		c.logger.Debug("Stored link preview image in S3", "asset_id", assetID, "size", len(data))
-		return asset, nil
-	}
-
-	objectKey := PublicServerAssetObjectKey(assetID)
-	meta := jetstream.ObjectMeta{
-		Name: objectKey,
-		Headers: map[string][]string{
-			"Content-Type": {contentType},
-		},
-	}
-	info, err := c.storage.serverAssets.Put(ctx, meta, bytes.NewReader(data))
-	if err != nil {
-		return nil, fmt.Errorf("upload link preview image to SERVER_ASSETS: %w", err)
-	}
-	asset.Size = int64(info.Size)
-	asset.Storage = &corev1.AssetRecord_Nats{Nats: &corev1.NATSAsset{Key: objectKey}}
-	c.logger.Debug("Stored link preview image in SERVER_ASSETS", "asset_id", assetID, "size", len(data))
-	return asset, nil
-}
-
-// ServerAssetInfo contains metadata about a server asset.
-type ServerAssetInfo struct {
-	Size        int64
-	ContentType string
-}
-
-// PublicServerAssetLocation binds a successful public classification to one
-// exact backend object. Its fields stay private so callers cannot manufacture
-// a location that bypasses ResolvePublicServerAsset.
-type PublicServerAssetLocation struct {
-	natsKey    string
-	natsNUID   string
-	natsDigest string
-	s3Key      string
-}
-
-func publicNATSServerAssetLocation(key string, info *jetstream.ObjectInfo) (*PublicServerAssetLocation, bool) {
-	if key == "" || info == nil || info.NUID == "" || info.Digest == "" {
-		return nil, false
-	}
-	return &PublicServerAssetLocation{
-		natsKey:    key,
-		natsNUID:   info.NUID,
-		natsDigest: info.Digest,
-	}, true
-}
-
-func serverAssetVisibilityMarkerMatches(info *jetstream.ObjectInfo) bool {
-	return info != nil && info.NUID != "" && info.Digest != "" &&
-		info.Headers.Get(ServerAssetVisibilityHeader) == ServerAssetVisibilityPublic &&
-		info.Headers.Get(ServerAssetVisibilityNUIDHeader) == info.NUID &&
-		info.Headers.Get(ServerAssetVisibilityDigestHeader) == info.Digest
-}
-
-// serverAssetRequestKey recognizes the explicit public/ namespace used by new
-// NATS objects and the flat canonical IDs used by historical NATS objects and
-// current S3 instance/ objects. Every other namespace fails closed.
-func serverAssetRequestKey(key string) (assetID string, namespaced bool, ok bool) {
-	if key == "" || key != strings.TrimSpace(key) || strings.HasPrefix(key, "/") || strings.Contains(key, "\\") {
-		return "", false, false
-	}
-	if strings.HasPrefix(key, PublicServerAssetObjectPrefix) {
-		namespaced = true
-		assetID = strings.TrimPrefix(key, PublicServerAssetObjectPrefix)
-	} else {
-		assetID = key
-	}
-	if len(assetID) != idLength+1 || assetID[0] != 'A' || strings.Contains(assetID, "/") {
-		return "", false, false
-	}
-	for _, ch := range assetID[1:] {
-		if !strings.ContainsRune(idAlphabet, ch) {
-			return "", false, false
-		}
-	}
-	return assetID, namespaced, true
-}
-
-func serverAssetNATSObjectKeys(key string) (logicalID string, namespaced bool, objectKeys []string, ok bool) {
-	logicalID, namespaced, ok = serverAssetRequestKey(key)
-	if !ok {
-		return "", false, nil, false
-	}
-	if namespaced {
-		return logicalID, true, []string{key}, true
-	}
-	return logicalID, false, []string{PublicServerAssetObjectKey(logicalID), logicalID}, true
-}
-
-// IsReservedServerAssetKey rejects private, internal, and unknown namespaces
-// before public-route transform parsing or backend probing.
-func IsReservedServerAssetKey(key string) bool {
-	_, _, ok := serverAssetRequestKey(key)
-	return !ok
-}
-
-// ResolvePublicServerAsset positively classifies an object and binds the
-// decision to one exact backend key before the public route performs cache
-// access, content reads, or transforms. Unknown objects fail closed.
-func (c *ChattoCore) ResolvePublicServerAsset(ctx context.Context, key string) (*PublicServerAssetLocation, bool) {
-	assetID, namespaced, ok := serverAssetRequestKey(key)
-	if !ok || c.assetModel == nil || c.assetModel.assets.Projection() == nil {
-		return nil, false
-	}
-
-	// Durable room-scoped declarations take precedence over every public hint,
-	// including stale metadata or a colliding current public reference.
-	assetState := c.assetModel.AssetState(assetID)
-	if assetState.Creation != nil || assetState.Deleted {
-		return nil, false
-	}
-
-	// The public/ namespace is itself the positive declaration for new NATS
-	// objects. Metadata-only inspection still rejects an object misplaced there
-	// by a private writer before any content or derivative cache is opened.
-	if namespaced {
-		info, err := c.storage.serverAssets.GetInfo(ctx, key)
-		if err != nil || info == nil || info.Headers.Get("Room-Id") != "" || info.Headers.Get("Upload-Id") != "" {
-			return nil, false
-		}
-		return publicNATSServerAssetLocation(key, info)
-	}
-
-	// Canonical-ID URLs remain aliases for new namespaced objects. This keeps
-	// stored API references and clients that retained the logical ID working.
-	if info, err := c.storage.serverAssets.GetInfo(ctx, PublicServerAssetObjectKey(assetID)); err == nil && info != nil {
-		if info.Headers.Get("Room-Id") != "" || info.Headers.Get("Upload-Id") != "" {
-			return nil, false
-		}
-		return publicNATSServerAssetLocation(PublicServerAssetObjectKey(assetID), info)
-	}
-
-	// Object metadata is safe to inspect for classification; object content is
-	// not opened until this method has returned true.
-	var legacyNATSExists bool
-	if info, err := c.storage.serverAssets.GetInfo(ctx, assetID); err == nil && info != nil {
-		legacyNATSExists = true
-		if info.Headers.Get("Room-Id") != "" || info.Headers.Get("Upload-Id") != "" {
-			return nil, false
-		}
-		if serverAssetVisibilityMarkerMatches(info) {
-			return publicNATSServerAssetLocation(assetID, info)
-		}
-	}
-
-	// Historical public objects predate the explicit visibility header. Their
-	// durable/current public references provide the positive declaration.
-	legacyDeclaredPublic := c.userModel != nil && c.userModel.isPublicAvatarAsset(assetID)
-	if c.configModel != nil {
-		logo := c.configModel.serverBrandingAsset("logo")
-		banner := c.configModel.serverBrandingAsset("banner")
-		if assetRecordMatchesKey(logo, assetID) || assetRecordMatchesKey(banner, assetID) {
-			legacyDeclaredPublic = true
-		}
-	}
-	if assetState.PublicLinkPreview {
-		legacyDeclaredPublic = true
-	}
-	if legacyDeclaredPublic && legacyNATSExists {
-		info, err := c.storage.serverAssets.GetInfo(ctx, assetID)
-		if err != nil || info == nil || info.Headers.Get("Room-Id") != "" || info.Headers.Get("Upload-Id") != "" {
-			return nil, false
-		}
-		return publicNATSServerAssetLocation(assetID, info)
-	}
-
-	// S3 server assets live exclusively below instance/. Private current and
-	// historical attachments use attachments/ or spaces/*/attachments/.
-	if c.s3Client != nil {
-		s3Key := S3KeyServerAsset(assetID)
-		if _, err := c.s3Client.StatObject(ctx, s3Key); err == nil {
-			return &PublicServerAssetLocation{s3Key: s3Key}, true
-		}
-	}
-	return nil, false
-}
-
-// IsPublicServerAsset reports whether ResolvePublicServerAsset can bind the
-// request key to an explicitly public object. Prefer the resolver in delivery
-// paths so classification cannot fall through to a different backend object.
-func (c *ChattoCore) IsPublicServerAsset(ctx context.Context, key string) bool {
-	_, ok := c.ResolvePublicServerAsset(ctx, key)
-	return ok
-}
-
-// GetPublicServerAsset opens only the exact object previously classified by
-// ResolvePublicServerAsset. It never probes a fallback backend or key.
-func (c *ChattoCore) GetPublicServerAsset(ctx context.Context, location *PublicServerAssetLocation) (io.Reader, *ServerAssetInfo, error) {
-	if location == nil {
-		return nil, nil, jetstream.ErrObjectNotFound
-	}
-	if location.natsKey != "" {
-		obj, err := c.storage.serverAssets.Get(ctx, location.natsKey)
-		if err != nil {
-			return nil, nil, err
-		}
-		info, err := obj.Info()
-		if err != nil || info == nil || info.NUID != location.natsNUID || info.Digest != location.natsDigest {
-			_ = obj.Close()
-			return nil, nil, jetstream.ErrObjectNotFound
-		}
-		return obj, &ServerAssetInfo{
-			Size:        int64(info.Size),
-			ContentType: info.Headers.Get("Content-Type"),
-		}, nil
-	}
-	if location.s3Key != "" && c.s3Client != nil {
-		reader, info, err := c.s3Client.GetObject(ctx, location.s3Key)
-		if err != nil {
-			return nil, nil, err
-		}
-		return reader, &ServerAssetInfo{Size: info.Size, ContentType: info.ContentType}, nil
-	}
-	return nil, nil, jetstream.ErrObjectNotFound
-}
-
-// ServerAssetRecordFromAnyBackend builds an AssetRecord by probing the
-// server-asset backends. It is primarily for legacy ID-only server-scoped
-// assets that need to be rehydrated into richer metadata.
-func (c *ChattoCore) ServerAssetRecordFromAnyBackend(ctx context.Context, assetID, filename string) (*corev1.AssetRecord, error) {
-	logicalID, namespaced, natsKeys, ok := serverAssetNATSObjectKeys(assetID)
-	if !ok {
-		return nil, jetstream.ErrObjectNotFound
-	}
-	var natsErr error
-	for _, objectKey := range natsKeys {
-		obj, err := c.storage.serverAssets.Get(ctx, objectKey)
-		if err != nil {
-			natsErr = err
-			continue
-		}
-		if closer, ok := obj.(io.Closer); ok {
-			defer closer.Close()
-		}
-		info, _ := obj.Info()
-		contentType := info.Headers.Get("Content-Type")
-		if contentType == "" {
-			contentType = "application/octet-stream"
-		}
-		return &corev1.AssetRecord{
-			Id:          logicalID,
-			Filename:    filename,
-			ContentType: contentType,
-			Size:        int64(info.Size),
-			Storage:     &corev1.AssetRecord_Nats{Nats: &corev1.NATSAsset{Key: objectKey}},
-		}, nil
-	}
-
-	if c.s3Client != nil && !namespaced {
-		s3Info, s3Err := c.s3Client.StatObject(ctx, S3KeyServerAsset(logicalID))
-		if s3Err == nil {
-			contentType := s3Info.ContentType
-			if contentType == "" {
-				contentType = "application/octet-stream"
-			}
-			return &corev1.AssetRecord{
-				Id:          logicalID,
-				Filename:    filename,
-				ContentType: contentType,
-				Size:        s3Info.Size,
-				Storage: &corev1.AssetRecord_S3{S3: &corev1.S3Asset{
-					Key:    logicalID,
-					Bucket: proto.String(c.s3Client.Bucket()),
-				}},
-			}, nil
-		}
-		c.logger.Debug("Server asset record not found in either backend",
-			"asset_id", logicalID,
-			"nats_error", natsErr,
-			"s3_error", s3Err)
-	}
-
-	return nil, natsErr
-}
-
-// GetServerAssetFromAnyBackend retrieves a server asset by probing both NATS and S3 backends.
-// It tries the canonical SERVER_ASSETS NATS object store first, then S3.
-// Returns a reader for the asset content and metadata.
-// The caller is responsible for closing the reader if it implements io.Closer.
-func (c *ChattoCore) GetServerAssetFromAnyBackend(ctx context.Context, assetID string) (io.Reader, *ServerAssetInfo, error) {
-	logicalID, namespaced, natsKeys, ok := serverAssetNATSObjectKeys(assetID)
-	if !ok {
-		return nil, nil, jetstream.ErrObjectNotFound
-	}
-	var natsErr error
-	for _, objectKey := range natsKeys {
-		obj, err := c.storage.serverAssets.Get(ctx, objectKey)
-		if err != nil {
-			natsErr = err
-			continue
-		}
-		info, _ := obj.Info()
-		return obj, &ServerAssetInfo{
-			Size:        int64(info.Size),
-			ContentType: info.Headers.Get("Content-Type"),
-		}, nil
-	}
-
-	// If NATS failed and S3 is configured, try S3
-	if c.s3Client != nil && !namespaced {
-		s3Key := S3KeyServerAsset(logicalID)
-		reader, s3Info, s3Err := c.s3Client.GetObject(ctx, s3Key)
-		if s3Err == nil {
-			return reader, &ServerAssetInfo{
-				Size:        s3Info.Size,
-				ContentType: s3Info.ContentType,
-			}, nil
-		}
-		// Log S3 error but return the original NATS error
-		c.logger.Debug("Instance asset not found in either backend",
-			"asset_id", logicalID,
-			"nats_error", natsErr,
-			"s3_error", s3Err)
-	}
-
-	return nil, nil, natsErr
-}
-
-// CleanupAsset deletes an asset from the server object store.
-// Used to clean up orphaned assets when subsequent operations fail.
-func (c *ChattoCore) CleanupAsset(ctx context.Context, asset *corev1.DeprecatedAsset) {
-	if asset == nil {
-		return
-	}
-	if natsAsset := asset.GetNats(); natsAsset != nil {
-		if err := c.storage.serverAssets.Delete(ctx, natsAsset.Key); err != nil {
-			c.logger.Warn("Failed to clean up orphaned asset", "key", natsAsset.Key, "error", err)
-		} else {
-			c.logger.Info("Cleaned up orphaned asset", "key", natsAsset.Key)
-		}
-	}
-	if s3Asset := asset.GetS3(); s3Asset != nil && c.s3Client != nil {
-		s3Key := S3KeyServerAsset(s3Asset.Key)
-		if err := c.s3Client.DeleteObjectFromBucket(ctx, s3Asset.GetBucket(), s3Key); err != nil {
-			c.logger.Warn("Failed to clean up orphaned S3 asset", "asset_id", s3Asset.Key, "s3_key", s3Key, "error", err)
-		} else {
-			c.logger.Info("Cleaned up orphaned S3 asset", "asset_id", s3Asset.Key, "s3_key", s3Key)
-		}
-	}
-	c.deleteCachedResizesForServerAsset(ctx, assetIDFromAsset(asset), "orphaned asset", "")
-}
-
-// deleteAsset deletes a server asset from its storage backend (NATS or S3).
-// This is a helper for cleaning up old assets when they are replaced.
-// For S3, the assetID stored in S3Asset.Key is used to construct the full S3 path.
-// The assetType and ownerID are used for logging only.
-func (c *ChattoCore) deleteAsset(ctx context.Context, asset *corev1.DeprecatedAsset, assetType, ownerID string) {
-	if asset == nil {
-		return
-	}
-	if natsAsset := asset.GetNats(); natsAsset != nil {
-		if err := c.storage.serverAssets.Delete(ctx, natsAsset.Key); err != nil {
-			c.logger.Warn("Failed to delete old "+assetType, "owner_id", ownerID, "key", natsAsset.Key, "error", err)
-		} else {
-			c.logger.Info("Deleted old "+assetType, "owner_id", ownerID, "key", natsAsset.Key)
-		}
-	}
-	if s3Asset := asset.GetS3(); s3Asset != nil && c.s3Client != nil {
-		// S3Asset.Key stores just the assetID; construct the full S3 path
-		s3Key := S3KeyServerAsset(s3Asset.Key)
-		if err := c.s3Client.DeleteObjectFromBucket(ctx, s3Asset.GetBucket(), s3Key); err != nil {
-			c.logger.Warn("Failed to delete old S3 "+assetType, "owner_id", ownerID, "asset_id", s3Asset.Key, "s3_key", s3Key, "error", err)
-		} else {
-			c.logger.Info("Deleted old S3 "+assetType, "owner_id", ownerID, "asset_id", s3Asset.Key, "s3_key", s3Key)
-		}
-	}
-	c.deleteCachedResizesForServerAsset(ctx, assetIDFromAsset(asset), assetType, ownerID)
-}
-
-func (c *ChattoCore) deleteCachedResizesForServerAsset(ctx context.Context, assetID, assetType, ownerID string) {
-	assetKeys := []string{assetID}
-	if logicalID, namespaced, ok := serverAssetRequestKey(assetID); ok {
-		if namespaced {
-			assetKeys = append(assetKeys, logicalID)
-		} else {
-			assetKeys = append(assetKeys, PublicServerAssetObjectKey(logicalID))
-		}
-	}
-	deletedCount := 0
-	var cacheErr error
-	for _, assetKey := range assetKeys {
-		count, err := c.mediaModel.DeleteCachedResizesForServerAsset(ctx, assetKey)
-		deletedCount += count
-		cacheErr = errors.Join(cacheErr, err)
-	}
-	if cacheErr != nil {
-		c.logger.Warn("Failed to delete cached resizes for server asset",
-			"asset_id", assetID,
-			"asset_type", assetType,
-			"owner_id", ownerID,
-			"error", cacheErr)
-	} else if deletedCount > 0 {
-		c.logger.Debug("Deleted cached resizes for server asset",
-			"asset_id", assetID,
-			"asset_type", assetType,
-			"owner_id", ownerID,
-			"deleted_count", deletedCount)
-	}
 }
 
 // Ready checks if the core is fully initialized and current persistent resources are accessible.
@@ -1057,420 +344,4 @@ func (c *ChattoCore) Subscribe(ctx context.Context, subject string, handler nats
 	}
 
 	return sub, nil
-}
-
-// ============================================================================
-// Storage
-// ============================================================================
-
-// storage encapsulates JetStream resources used by Chatto Core.
-type storage struct {
-	encryptionKV   jetstream.KeyValue // ENCRYPTION_KEYS - KMS KEKs (excluded from backups)
-	runtimeStateKV jetstream.KeyValue // RUNTIME_STATE  - persisted latest-value runtime/user state + wrapped app DEKs
-
-	serverAssets    jetstream.ObjectStore // SERVER_ASSETS - all NATS-backed asset binaries
-	serverEvtStream jetstream.Stream      // EVT       - event-sourcing log (ADR-033/034).
-
-	memoryCacheKV   jetstream.KeyValue    // MEMORY_CACHE - volatile, memory-backed runtime cache state
-	imageCacheStore jetstream.ObjectStore // Optional: cached resized images (nil if disabled)
-}
-
-// newStorage initializes current JetStream resources.
-func newStorage(js jetstream.JetStream, ctx context.Context, cfg config.CoreConfig) (*storage, error) {
-	// Initialize KMS KEK bucket (excluded from backups for security). App-owned
-	// wrapped DEK records live in RUNTIME_STATE so normal backups keep encrypted
-	// content together with its wrapped content-key registry, but not the KEKs
-	// needed to unwrap it.
-	encryptionKV, err := createJetStreamResourceWithRetry(ctx, func(ctx context.Context) (jetstream.KeyValue, error) {
-		return js.CreateOrUpdateKeyValue(ctx, jetstream.KeyValueConfig{
-			Bucket:      "ENCRYPTION_KEYS",
-			Description: "KMS key-encryption keys (excluded from backups)",
-			Storage:     jetstream.FileStorage,
-			History:     1,
-			Replicas:    cfg.Replicas,
-		})
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create ENCRYPTION_KEYS KV bucket: %w", err)
-	}
-
-	runtimeStateKV, err := createJetStreamResourceWithRetry(ctx, func(ctx context.Context) (jetstream.KeyValue, error) {
-		return js.CreateOrUpdateKeyValue(ctx, jetstream.KeyValueConfig{
-			Bucket:         "RUNTIME_STATE",
-			Description:    "Persisted latest-value runtime/user state",
-			Storage:        jetstream.FileStorage,
-			History:        1,
-			Compression:    true,
-			Replicas:       cfg.Replicas,
-			LimitMarkerTTL: 24 * time.Hour,
-		})
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create RUNTIME_STATE KV bucket: %w", err)
-	}
-
-	memoryCacheKV, err := createJetStreamResourceWithRetry(ctx, func(ctx context.Context) (jetstream.KeyValue, error) {
-		return js.CreateOrUpdateKeyValue(ctx, jetstream.KeyValueConfig{
-			Bucket:         "MEMORY_CACHE",
-			Description:    "Volatile memory-backed runtime cache state",
-			Storage:        jetstream.MemoryStorage,
-			History:        1,
-			Replicas:       cfg.Replicas,
-			LimitMarkerTTL: PresenceTTL,
-		})
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create MEMORY_CACHE KV bucket: %w", err)
-	}
-
-	// Initialize image cache object store (optional, only when enabled)
-	var imageCacheStore jetstream.ObjectStore
-	if cfg.Assets.Cache.Enabled {
-		imageCacheStore, err = createJetStreamResourceWithRetry(ctx, func(ctx context.Context) (jetstream.ObjectStore, error) {
-			return js.CreateOrUpdateObjectStore(ctx, jetstream.ObjectStoreConfig{
-				Bucket:      "ASSET_CACHE",
-				Description: "Cached resized images",
-				Storage:     jetstream.FileStorage,
-				Compression: true,
-				TTL:         cfg.Assets.Cache.TTLOrDefault(),
-				Replicas:    cfg.Replicas,
-			})
-		})
-		if err != nil {
-			return nil, fmt.Errorf("failed to create ASSET_CACHE object store: %w", err)
-		}
-	}
-
-	serverAssets, err := createJetStreamResourceWithRetry(ctx, func(ctx context.Context) (jetstream.ObjectStore, error) {
-		return js.CreateOrUpdateObjectStore(ctx, jetstream.ObjectStoreConfig{
-			Bucket:      "SERVER_ASSETS",
-			Description: "Server asset binaries (avatars, branding, link previews, attachments)",
-			Storage:     jetstream.FileStorage,
-			Compression: true,
-			Replicas:    cfg.Replicas,
-		})
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create SERVER_ASSETS object store: %w", err)
-	}
-
-	// EVT — the event-sourcing log (ADR-033/034).
-	// Subjects are evt.{aggregateType}.{aggregateId}.{eventType}; live.evt.> is
-	// the republish target so projections and live subscribers consume
-	// from a single NATS Core path.
-	evtMetadata, err := prepareEVTStreamMetadata(ctx, js)
-	if err != nil {
-		return nil, fmt.Errorf("prepare EVT stream metadata: %w", err)
-	}
-	evtConfig := jetstream.StreamConfig{
-		Name:        "EVT",
-		Description: "Event-sourcing log (ADR-033)",
-		Subjects:    []string{"evt.>"},
-		Storage:     jetstream.FileStorage,
-		Compression: jetstream.S2Compression,
-		Replicas:    cfg.Replicas,
-		Metadata:    evtMetadata,
-		// AllowAtomicPublish gates the Nats-Batch-Id / Nats-Batch-Commit
-		// protocol on this stream. Used by Publisher.AppendBatch to
-		// land multi-aggregate cascades (MoveRoomToGroup, DM creation)
-		// adjacently in stream order so projections never observe an
-		// intermediate state that breaks an invariant.
-		AllowAtomicPublish: true,
-		RePublish: &jetstream.RePublish{
-			Source:      "evt.>",
-			Destination: "live.evt.>",
-		},
-	}
-	serverEvtStream, err := createJetStreamResourceWithRetry(ctx, func(ctx context.Context) (jetstream.Stream, error) {
-		return js.CreateOrUpdateStream(ctx, evtConfig)
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to create EVT stream: %w", err)
-	}
-	if !evtstream.ValidIdentity(evtConfig.Metadata[evtstream.IdentityMetadataKey]) {
-		info := serverEvtStream.CachedInfo()
-		if info == nil {
-			return nil, fmt.Errorf("created EVT stream info is unavailable")
-		}
-		identity, identityErr := evtstream.NewIdentity(info.Created)
-		if identityErr != nil {
-			return nil, identityErr
-		}
-		evtConfig.Metadata[evtstream.IdentityMetadataKey] = identity
-		serverEvtStream, err = createJetStreamResourceWithRetry(ctx, func(ctx context.Context) (jetstream.Stream, error) {
-			return js.CreateOrUpdateStream(ctx, evtConfig)
-		})
-		if err != nil {
-			return nil, fmt.Errorf("persist EVT stream identity: %w", err)
-		}
-	}
-
-	return &storage{
-		encryptionKV:    encryptionKV,
-		runtimeStateKV:  runtimeStateKV,
-		serverAssets:    serverAssets,
-		serverEvtStream: serverEvtStream,
-		memoryCacheKV:   memoryCacheKV,
-		imageCacheStore: imageCacheStore,
-	}, nil
-}
-
-func prepareEVTStreamMetadata(ctx context.Context, js jetstream.JetStream) (map[string]string, error) {
-	metadata := make(map[string]string)
-	stream, err := js.Stream(ctx, "EVT")
-	switch {
-	case err == nil:
-		info, infoErr := stream.Info(ctx)
-		if infoErr != nil {
-			return nil, fmt.Errorf("read existing EVT stream info: %w", infoErr)
-		}
-		for key, value := range info.Config.Metadata {
-			metadata[key] = value
-		}
-	case errors.Is(err, jetstream.ErrStreamNotFound):
-	case err != nil:
-		return nil, fmt.Errorf("open existing EVT stream: %w", err)
-	}
-	if evtstream.ValidIdentity(metadata[evtstream.IdentityMetadataKey]) {
-		return metadata, nil
-	}
-	if stream == nil {
-		return metadata, nil
-	}
-	if stream.CachedInfo() == nil {
-		return nil, fmt.Errorf("existing EVT stream info is unavailable")
-	}
-	identity, err := evtstream.NewIdentity(stream.CachedInfo().Created)
-	if err != nil {
-		return nil, err
-	}
-	metadata[evtstream.IdentityMetadataKey] = identity
-	return metadata, nil
-}
-
-func createJetStreamResourceWithRetry[T any](ctx context.Context, create func(context.Context) (T, error)) (T, error) {
-	const maxAttempts = 3
-
-	var zero T
-	for attempt := 1; attempt <= maxAttempts; attempt++ {
-		resource, err := create(ctx)
-		if err == nil {
-			return resource, nil
-		}
-		if attempt == maxAttempts || !isTransientJetStreamStoreCreateError(err) {
-			return zero, err
-		}
-
-		timer := time.NewTimer(time.Duration(attempt) * 25 * time.Millisecond)
-		select {
-		case <-ctx.Done():
-			timer.Stop()
-			return zero, ctx.Err()
-		case <-timer.C:
-		}
-	}
-
-	return zero, nil
-}
-
-func isTransientJetStreamStoreCreateError(err error) bool {
-	type apiErrorProvider interface {
-		APIError() *jetstream.APIError
-	}
-
-	var provider apiErrorProvider
-	if !errors.As(err, &provider) {
-		return false
-	}
-	apiErr := provider.APIError()
-	if apiErr == nil {
-		return false
-	}
-	return (apiErr.ErrorCode == 10049 && strings.Contains(apiErr.Description, "error creating store for stream")) ||
-		(apiErr.ErrorCode == 10058 && strings.Contains(apiErr.Description, "stream name already in use"))
-}
-
-// ============================================================================
-// KV Key Helpers
-// ============================================================================
-
-// These helper functions format keys for NATS KV bucket entries. They stay in
-// the core package since they're only used here and are integral to how core
-// interacts with storage.
-
-// userKey returns the KV key for a user record.
-func userKey(userID string) string {
-	return fmt.Sprintf("user.%s", userID)
-}
-
-// userByLoginKey returns the KV key for a login-to-userID index entry.
-// Login names are lowercase to ensure case-insensitive lookups.
-func userByLoginKey(login string) string {
-	return fmt.Sprintf("user_by_login.%s", strings.ToLower(login))
-}
-
-// userAuthPasswordKey returns the KV key for a user's password hash.
-// This follows the pattern auth.{userId}.{method}.{field} for future extensibility.
-func userAuthPasswordKey(userID string) string {
-	return fmt.Sprintf("auth.%s.password", userID)
-}
-
-// userAvatarKey returns the KV key for a user's avatar asset reference.
-// Avatar assets are stored separately from user profile to avoid overwriting
-// the entire user record when the avatar changes.
-func userAvatarKey(userID string) string {
-	return fmt.Sprintf("user.%s.avatar", userID)
-}
-
-// roomKey returns the KV key for a room record in a space bucket.
-// Pattern: `room.{kind}.{roomID}` where kind is "channel" or "dm".
-func roomKey(kind RoomKind, roomID string) string {
-	return fmt.Sprintf("room.%s.%s", kind, roomID)
-}
-
-// roomKeyPrefix returns the key prefix for listing all rooms of a given
-// kind in a CONFIG bucket. Pattern: `room.{kind}.*`.
-func roomKeyPrefix(kind RoomKind) string {
-	return fmt.Sprintf("room.%s.*", kind)
-}
-
-// roomNameIndexKey returns the KV key that claims a room name within a space.
-// Names are lowercased and trimmed so the claim is case-insensitive. The value
-// stored at this key is the room ID, which lets us recover from partial failures
-// (a stale claim whose room never got written can be reclaimed by the same room
-// trying again).
-func roomNameIndexKey(name string) string {
-	return fmt.Sprintf("room_name_index.%s", strings.ToLower(strings.TrimSpace(name)))
-}
-
-// ============================================================================
-// Event Publishing Helpers
-// ============================================================================
-
-// natsPublishFlushTimeout bounds how long a fire-and-forget publish will wait
-// for the NATS server to acknowledge buffered bytes. Without a timeout, a
-// hung server (e.g. network partition) would block the calling goroutine
-// indefinitely instead of surfacing as a normal error.
-const natsPublishFlushTimeout = 5 * time.Second
-
-// publishLiveEvent publishes a transient LiveEvent directly to a live.sync.>
-// subject, bypassing JetStream storage. The subject should already include
-// the "live.sync." prefix.
-func (c *ChattoCore) publishLiveEvent(_ context.Context, subject string, event *corev1.LiveEvent) error {
-	if err := validateLiveEvent(event); err != nil {
-		return err
-	}
-
-	eventData, err := proto.Marshal(event)
-	if err != nil {
-		return fmt.Errorf("failed to marshal live event: %w", err)
-	}
-
-	if err := c.nc.Publish(subject, eventData); err != nil {
-		return fmt.Errorf("failed to publish live event to %s: %w", subject, err)
-	}
-
-	if err := c.nc.FlushTimeout(natsPublishFlushTimeout); err != nil {
-		return fmt.Errorf("failed to flush live event to %s: %w", subject, err)
-	}
-	return nil
-}
-
-func validateEvent(event *corev1.Event) error {
-	if event == nil || event.Event == nil {
-		return fmt.Errorf("%w: event payload is nil or oneof field is unset", ErrInvalidEvent)
-	}
-	return nil
-}
-
-func validateLiveEvent(event *corev1.LiveEvent) error {
-	if event == nil || event.Event == nil {
-		return fmt.Errorf("%w: live event payload is nil or oneof field is unset", ErrInvalidEvent)
-	}
-	return nil
-}
-
-// newEvent fills in the Id, ActorID, and CreatedAt fields of an Event
-// envelope if they're not already set. The caller provides the event
-// with the concrete oneof variant already populated.
-func newEvent(actorID string, event *corev1.Event) *corev1.Event {
-	if event.Id == "" {
-		event.Id = NewEventID()
-	}
-	if event.ActorId == "" {
-		event.ActorId = actorID
-	}
-	if event.CreatedAt == nil {
-		event.CreatedAt = timestamppb.New(time.Now())
-	}
-	return event
-}
-
-// newLiveEvent fills in the Id, ActorID, and CreatedAt fields of a LiveEvent
-// envelope if they're not already set. The caller provides the event with the
-// concrete oneof variant already populated.
-func newLiveEvent(actorID string, event *corev1.LiveEvent) *corev1.LiveEvent {
-	if event.Id == "" {
-		event.Id = NewEventID()
-	}
-	if event.ActorId == "" {
-		event.ActorId = actorID
-	}
-	if event.CreatedAt == nil {
-		event.CreatedAt = timestamppb.New(time.Now())
-	}
-	return event
-}
-
-// ============================================================================
-// Event Streaming
-// ============================================================================
-
-// isTerminalIteratorError returns true if the error indicates the iterator
-// cannot be recovered (connection closed, consumer deleted, etc.).
-// Recoverable errors (heartbeat missed, leadership changed) return false.
-func isTerminalIteratorError(err error) bool {
-	if err == nil {
-		return false
-	}
-	// Terminal errors - cannot recover, must stop
-	if errors.Is(err, jetstream.ErrMsgIteratorClosed) ||
-		errors.Is(err, jetstream.ErrConnectionClosed) ||
-		errors.Is(err, jetstream.ErrServerShutdown) ||
-		errors.Is(err, jetstream.ErrConsumerDeleted) {
-		return true
-	}
-	return false
-}
-
-// ============================================================================
-// Statistics
-// ============================================================================
-
-// ServerStats contains aggregate counts surfaced in the admin dashboard.
-type ServerStats struct {
-	UserCount        int
-	ChannelRoomCount int
-	DMRoomCount      int
-}
-
-// GetStats returns deployment-level counts: registered users, channel rooms,
-// DM rooms. Per-space breakdowns went away with the Space tier (ADR-030).
-func (c *ChattoCore) GetStats(ctx context.Context) (*ServerStats, error) {
-	stats := &ServerStats{}
-	stats.UserCount = c.userModel.userCount()
-
-	channelRooms, err := c.ListRooms(ctx, KindChannel)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list channel rooms: %w", err)
-	}
-	stats.ChannelRoomCount = len(channelRooms)
-
-	dmRooms, err := c.ListRooms(ctx, KindDM)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list dm rooms: %w", err)
-	}
-	stats.DMRoomCount = len(dmRooms)
-
-	return stats, nil
 }
